+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file - [read more
 ### Fixed
 - Symfony 4.2 traces generation #280
 - Drupal crashes (temporary workaround) #285
+- Tracing of http status code in generic web requests #288
 
 ## [0.12.2]